backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
requests==2.31.0
tenacity==8.2.3